# apps/api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "admin"
                ],
                "summary": "Create service",
                "parameters": [
                    {
                        "description": "Service payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateServiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/registry.Service"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "admin"
                ],
                "summary": "Create service",
                "parameters": [
                    {
                        "description": "Service payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateServiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/registry.Service"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.CreateServiceRequest": {
            "type": "object",
            "properties": {
                "base_url": {
                    "type": "string",
                    "example": "http://user-service:8081"
                },
                "description": {
                    "type": "string",
                    "example": "Manages users"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "User Service"
                },
                "public_prefix": {
                    "type": "string",
                    "example": "/api/users/"
                },
                "swagger_url": {
                    "type": "string",
                    "example": "http://user-service:8081/swagger.json"
                }
            }
        },
        "registry.Service": {
            "type": "object",
            "properties": {
                "base_url": {
                    "type": "string",
                    "example": "http://user-service:8081"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-11-22T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Manages users"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "string",
                    "example": "3d1a7e94-0a2f-4a49-9a9b-8f9f2d0c6f67"
                },
                "last_health_at": {
                    "type": "string",
                    "example": "2025-11-22T10:20:00Z"
                },
                "last_refreshed_at": {
                    "type": "string",
                    "example": "2025-11-22T10:20:30Z"
                },
                "last_status": {
                    "type": "string",
                    "example": "Healthy"
                },
                "name": {
                    "type": "string",
                    "example": "User Service"
                },
                "public_prefix": {
                    "type": "string",
                    "example": "/api/users/"
                },
                "swagger_json": {},
                "swagger_url": {
                    "type": "string",
                    "example": "http://user-service:8081/swagger.json"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-11-22T10:10:00Z"
                }
            }
        }
    }
}